Replaced Emeraldite Shards with a Custom Item
Emerald Ingots are used to create emerald gear instead of emeraldite shards.
they require tin and an emerald block to create.
Why tin? It is the overworld ore with the least amount of recipes made from it


Emeraldite Ore no longer spawns, and the recipe for the shards has been changed back to emeralds.