"Fixed" Tier 6 Champions enchanted books having their bonus enchantment applied instead of a stored enchantment
Hard to explain in a short title, basically enchanted books store their enchantments in a different part of their nbt so the enchantments don't actually work when attacking with the book, and so the player can apply it to equipment. However the book was getting the mending enchantments applied to it directly instead so it couldn't be applied to equipment without disenchanting it first.

Now tier 5/6 champions drop apotheosis tomes instead, right-clicking with them turns them into an enchanted book. This also has the benefit of similar enchantments appearing on the same book, like the weapons tome will only have weapon enchantments, meaning splitting the book is less likely to be needed.

This also madHard to explain in a short title, basically enchanted books store their enchantments in a different part of their nbt so the enchantments don't actually work when attacking with the book, and so the player can apply it to equipment. However the book was getting the mending enchantments applied to it directly instead so it couldn't be applied to equipment without disenchanting it first.

Now tier 5/6 champions drop apotheosis tomes instead, right-clicking with them turns them into an enchanted book. This also has the benefit of similar enchantments appearing on the same book, like the weapons tome will only have weapon enchantments, meaning splitting the book is less likely to be needed.

Also added a new book drop to tier 4 champions