Enforced Protection & Safeguard Incompatibilities
Having Protection and Safeguard on the same gear (through exploits and whatnot, they are normally incompatible) will forcibly remove Protection off of that piece.
Warning: this is done in a very scuffed way and unexpected things could happen, so don't try to put those two enchants on the same piece!